backend/requirements.txt
openai
requests
python-dotenv
fastapi
uvicorn
pydantic